[add] New examples; readme expanded [ref] split to few files
